creating delete and put api

+++ src/main/java/com/effigoLearn/Learning/website/repo/StudentRepo.java
@@ -3,6 +3,6 @@ package com.effigoLearn.Learning.website.repo;
 import com.effigoLearn.Learning.website.model.StudentModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepo extends JpaRepository<StudentModel, Integer> {
+public interface StudentRepo extends JpaRepository<StudentModel, Long> {
 
 }
